Split UI into Report and Widgets
