+++ lib/Echidna/UI.hs
@@ -1,6 +1,5 @@
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE LambdaCase #-}
-{-# LANGUAGE NamedFieldPuns #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TupleSections #-}
 {-# LANGUAGE TemplateHaskell #-}
@@ -10,42 +9,29 @@ module Echidna.UI where
 
 import Brick
 import Brick.BChan
-import Brick.Widgets.Border
-import Brick.Widgets.Center
 import Control.Lens
 import Control.Monad (liftM2, liftM3, void, when)
 import Control.Monad.Catch (MonadCatch(..))
 import Control.Monad.IO.Class (MonadIO(..))
 import Control.Monad.Reader (MonadReader, runReader)
 import Control.Monad.Random.Strict (MonadRandom)
-import Data.Either (either)
 import Data.Has (Has(..))
-import Data.List (nub, intersperse)
-import Data.Map (Map)
-import Data.Maybe (catMaybes, maybe, fromMaybe)
-import Data.Set (Set)
-import Data.Version (showVersion)
+import Data.Maybe (maybe, fromMaybe)
 import EVM (VM)
-import EVM.Types (Addr, W256)
 import Graphics.Vty (Event(..), Key(..), Modifier(..), defaultConfig, mkVty)
 import System.Posix.Terminal (queryTerminal)
 import System.Posix.Types (Fd(..))
 import System.Timeout (timeout)
 import UnliftIO (MonadUnliftIO)
 import UnliftIO.Concurrent (killThread, forkIO)
-import Text.Printf (printf)
-
-import qualified Brick.AttrMap as A
-import qualified Data.Text as T
-import qualified Graphics.Vty as V
-import qualified Paths_echidna (version)
 
 import Echidna.Campaign
 import Echidna.ABI
-import Echidna.Exec
 import Echidna.Solidity
 import Echidna.Test
 import Echidna.Transaction
+import Echidna.UI.Report
+import Echidna.UI.Widgets
 
 data UIConf = UIConf { _dashboard :: Bool
                      , _maxTime   :: Maybe Int
@@ -53,153 +39,6 @@ data UIConf = UIConf { _dashboard :: Bool
                      }
 
 makeLenses ''UIConf
-
-data UIState = Uninitialized | Running | Timedout
-
--- | An address involved with a 'Transaction' is either the sender, the recipient, or neither of those things.
-data Role = Sender | Receiver | Ambiguous
-
--- | Rules for pretty-printing addresses based on their role in a transaction.
-type Names = Role -> Addr -> String
-
--- | Given rules for pretty-printing associated address, and whether to print them, pretty-print a 'Transaction'.
-ppTx :: (MonadReader x m, Has Names x, Has TxConf x) => Bool -> Tx -> m String
-ppTx pn (Tx c s r g gp v (t, b)) = let sOf = either ppSolCall (const "<CREATE>") in do
-  names <- view hasLens
-  tGas  <- view $ hasLens . txGas
-  return $ sOf c ++ (if not pn    then "" else names Sender s ++ names Receiver r)
-                 ++ (if g == tGas then "" else " Gas: "         ++ show g)
-                 ++ (if gp == 0   then "" else " Gas price: "   ++ show gp)
-                 ++ (if v == 0    then "" else " Value: "       ++ show v)
-                 ++ (if t == 0    then "" else " Time delay: "  ++ show t)
-                 ++ (if b == 0    then "" else " Block delay: " ++ show b)
-
--- | Given a number of boxes checked and a number of total boxes, pretty-print progress in box-checking.
-progress :: Int -> Int -> String
-progress n m = "(" ++ show n ++ "/" ++ show m ++ ")"
-
--- | Pretty-print the status of a solved test.
-ppFail :: (MonadReader x m, Has Names x, Has TxConf x) => Maybe (Int, Int) -> [Tx] -> m String
-ppFail _ [] = pure "failed with no transactions made ⁉️  "
-ppFail b xs = let status = case b of
-                                Nothing    -> ""
-                                Just (n,m) -> ", shrinking " ++ progress n m
-                  pxs = mapM (ppTx $ length (nub $ view src <$> xs) /= 1) xs in
- (("failed!💥  \n  Call sequence" ++ status ++ ":\n") ++) . unlines . fmap ("    " ++) <$> pxs
-
--- | Pretty-print the status of a test.
-ppTS :: (MonadReader x m, Has CampaignConf x, Has Names x, Has TxConf x) => TestState -> m String
-ppTS (Failed e)  = pure $ "could not evaluate ☣\n  " ++ show e
-ppTS (Solved l)  = ppFail Nothing l
-ppTS Passed      = pure "passed! 🎉"
-ppTS (Open i)    = view hasLens >>= \(CampaignConf t _ _ _ _ _ _) ->
-                     if i >= t then ppTS Passed else pure $ "fuzzing " ++ progress i t
-ppTS (Large n l) = view (hasLens . to shrinkLimit) >>= \m -> ppFail (if n < m then Just (n,m)
-                                                                              else Nothing) l
-
--- | Pretty-print the status of all 'SolTest's in a 'Campaign'.
-ppTests :: (MonadReader x m, Has CampaignConf x, Has Names x, Has TxConf x) => Campaign -> m String
-ppTests (Campaign ts _ _) = unlines . catMaybes <$> mapM pp ts where
-  pp (Left  (n, _), s)      = Just .                    ((T.unpack n ++ ": ") ++) <$> ppTS s
-  pp (Right _,      Open _) = pure Nothing
-  pp (Right (n, _), s)      = Just . (("assertion in " ++ T.unpack n ++ ": ") ++) <$> ppTS s
-
--- | Pretty-print the coverage a 'Campaign' has obtained.
-ppCoverage :: Map W256 (Set Int) -> Maybe String
-ppCoverage s | s == mempty = Nothing
-             | otherwise   = Just $ "Unique instructions: " ++ show (coveragePoints s)
-                                 ++ "\nUnique codehashes: " ++ show (length s)
-
-ppCampaign :: (MonadReader x m, Has CampaignConf x, Has Names x, Has TxConf x) => Campaign -> m String
-ppCampaign c = (++) <$> ppTests c <*> pure (maybe "" ("\n" ++) . ppCoverage $ c ^. coverage)
-
--- | Render 'Campaign' progress as a 'Widget'.
-campaignStatus :: (MonadReader x m, Has CampaignConf x, Has Names x, Has TxConf x)
-               => (Campaign, UIState) -> m (Widget ())
-campaignStatus (c@Campaign{_tests, _coverage}, uiState) = do
-  done <- isDone c
-  case (uiState, done) of
-    (Uninitialized, _) -> pure $ mainbox (padLeft (Pad 1) $ str "Starting up, please wait...") emptyWidget
-    (Timedout, _)      -> mainbox <$> testsWidget <*> pure (str "Timed out, C-c or esc to print report")
-    (_, True)          -> mainbox <$> testsWidget <*> pure (str "Campaign complete, C-c or esc to print report")
-    _                  -> mainbox <$> testsWidget <*> pure emptyWidget
-  where
-  mainbox :: Widget () -> Widget () -> Widget ()
-  mainbox inner underneath =
-    padTop (Pad 1) $ hCenter $ hLimit 120 $
-    (
-      borderWithLabel (withAttr "bold" $ str title) $
-      summaryWidget
-      <=>
-      hBorderWithLabel (str "Tests")
-      <=>
-      inner
-    )
-    <=>
-    hCenter underneath
-
-  title = "Echidna " ++ showVersion Paths_echidna.version
-
-  summaryWidget =
-    padLeft (Pad 1) (
-      str ("Tests found: " ++ show (length _tests))
-      <=>
-      maybe emptyWidget str (ppCoverage _coverage)
-    )
-
-  testsWidget :: (MonadReader x m, Has CampaignConf x, Has Names x, Has TxConf x)
-              => m (Widget())
-  testsWidget = foldl (<=>) emptyWidget . intersperse hBorder <$> traverse testWidget _tests
-
-  testWidget :: (MonadReader x m, Has CampaignConf x, Has Names x, Has TxConf x)
-             => (SolTest, TestState) -> m (Widget ())
-  testWidget (test, testState) =
-    case test of
-        Left  (n, _) -> widget n ""
-        Right (n, _) -> widget n "assertion in "
-    where
-    widget n infront = do
-      (status, details) <- tsWidget testState
-      pure $ padLeft (Pad 1) $
-        str infront <+> name n <+> str ": " <+> status
-        <=> padTop (Pad 1) details
-    name n = withAttr "bold" $ str (T.unpack n)
-
-  tsWidget :: (MonadReader x m, Has CampaignConf x, Has Names x, Has TxConf x)
-           => TestState -> m (Widget (), Widget ())
-  tsWidget (Failed e)  = pure (str "could not evaluate", str $ show e)
-  tsWidget (Solved l)  = failWidget Nothing l
-  tsWidget Passed      = pure (withAttr "success" $ str "PASSED!", emptyWidget)
-  tsWidget (Open i)    = view hasLens >>= \(CampaignConf t _ _ _ _ _ _) ->
-    if i >= t then
-      tsWidget Passed
-    else
-      pure (withAttr "working" $ str $ "fuzzing " ++ progress i t, emptyWidget)
-  tsWidget (Large n l) = view (hasLens . to shrinkLimit) >>= \m ->
-    failWidget (if n < m then Just (n,m) else Nothing) l
-
-  failWidget :: (MonadReader x m, Has Names x, Has TxConf x)
-             => Maybe (Int, Int) -> [Tx] -> m (Widget (), Widget ())
-  failWidget _ [] = pure (failureBadge, str "*no transactions made*")
-  failWidget b xs = do
-    s <- seqWidget
-    pure (failureBadge, titleWidget <=> s)
-    where
-    titleWidget  = str "Call sequence" <+> status <+> str ":"
-
-    status = case b of
-      Nothing    -> emptyWidget
-      Just (n,m) -> str ", " <+> withAttr "working" (str ("shrinking " ++ progress n m))
-
-    seqWidget = do
-      ppTxs <- mapM (ppTx $ length (nub $ view src <$> xs) /= 1) xs
-      let ordinals = str . printf "%d." <$> [1 :: Int ..]
-      pure $
-        foldl (<=>) emptyWidget $
-          zipWith (<+>) ordinals (withAttr "tx" . str <$> ppTxs)
-
-  failureBadge :: Widget ()
-  failureBadge = withAttr "failure" $ str "FAILED!"
 
 -- | Check if we should stop drawing (or updating) the dashboard, then do the right thing.
 monitor :: (MonadReader x m, Has CampaignConf x, Has Names x, Has TxConf x)
@@ -212,14 +51,6 @@ monitor = let
   se c (VtyEvent (EvKey KEsc _))                         = halt c
   se c (VtyEvent (EvKey (KChar 'c') l)) | MCtrl `elem` l = halt c
   se c _                                                 = continue c
-
-  attrs = A.attrMap (V.white `on` V.black)
-    [ ("failure", fg V.brightRed)
-    , ("bold", fg V.white `V.withStyle` V.bold)
-    , ("tx", fg V.brightWhite)
-    , ("working", fg V.brightBlue)
-    , ("success", fg V.brightGreen)
-    ]
   in
     liftM3 (,,) (view hasLens) (view hasLens) (view hasLens) <&> \s ->
       App (pure . cs s) neverShowCursor se pure (const attrs)
